% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{add_chars}
\alias{add_chars}
\title{Add Characters After Specific Number}
\usage{
add_chars(line, cols, chars = "|")
}
\arguments{
\item{line}{character of length 1}

\item{cols}{characters preceding those to add}

\item{chars}{characters to add, length 1 or same as cols}
}
\value{
character, col breaks added after cols

pdf2tab:::add_chars("ABcdEF", c(2, 4))
pdf2tab:::add_chars("ABcdEF", c(2, 4), 1:2)
}
\description{
Add Characters After Specific Number
}
