% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{add_print_rows}
\alias{add_print_rows}
\title{Add Rows to Text}
\usage{
add_print_rows(x, rows, cols)
}
\arguments{
\item{x}{list of characters}

\item{rows}{where to add row breaks}

\item{cols}{column breaks}
}
\value{
x with row break character strings added

pdf2tab:::add_rows(list("AB|CD|EF", "AB|CD|EF", "AB|CD|EF", "12|34|56"),
rows = c(3, 4))
}
\description{
Add Rows to Text
}
