% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{add_rows_recur}
\alias{add_rows_recur}
\title{Recursively add Rows}
\usage{
add_rows_recur(x, rows, new_row)
}
\arguments{
\item{x}{list of characters}

\item{rows}{where to add}

\item{new_row}{character, value for new row}
}
\value{
x with new row added after rows

pdf2tab:::add_rows_recur(list("ABC", "ABC"), NULL, "")
pdf2tab:::add_rows_recur(list("ABC", "ABC"), 1, "---")
}
\description{
Recursively add Rows
}
