% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pdf_load_funs.R,
%   R/pdf_page_funs.R
\name{drop_lines}
\alias{drop_lines}
\alias{drop_lines.pdf}
\alias{drop_lines.pdf_page}
\title{Drop lines from top and bottom of pages}
\usage{
drop_lines(x, lines, from)

\method{drop_lines}{pdf}(x, lines, from)

\method{drop_lines}{pdf_page}(x, lines, from)
}
\arguments{
\item{x}{an object of class pdf}

\item{lines}{lines to drop, either a number, or a vector of same length as number of pages in pdf}

\item{from}{either "top" or "bottom"}
}
\value{
pdf with lines dropped from each page
}
\description{
Drop lines from top and bottom of pages

Drop lines from pdf
}
\examples{
p <- pdf_load(test_pdf_text())
p <- drop_lines(p, 1, "top")
p <- drop_lines(p, 2, "bottom")
}
