% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{get_page_attr}
\alias{get_page_attr}
\title{Get Attribute from Page}
\usage{
get_page_attr(x, which)
}
\arguments{
\item{x}{pdf_age}

\item{which}{attribute}
}
\value{
attribute of x
}
\description{
Get Attribute from Page
}
\examples{
p <- pdf2tab:::pdf_page("ABC", 1, 2)
pdf2tab:::get_page_attr(p, "text")
}
