% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{get_page_cols}
\alias{get_page_cols}
\title{Get PDF Page Columns}
\usage{
get_page_cols(x)
}
\arguments{
\item{x}{pdf_age}
}
\value{
numeric, columns of x

p <- pdf2tab:::pdf_page("ABC", 1, 2)
pdf2tab:::get_page_cols(p)
}
\description{
Get PDF Page Columns
}
