% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_funs.R
\name{get_pdf_pages}
\alias{get_pdf_pages}
\title{Get Pages from PDF Object}
\usage{
get_pdf_pages(x, i = NULL)
}
\arguments{
\item{x}{pdf object}

\item{i}{which pages to get}
}
\value{
x <- pdf(test_pdf_text())
get_pdf_pages(x)
get_pdf_pages(x, 1:2)
}
\description{
Get Pages from PDF Object
}
