% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_load_funs.R
\name{keep_pages}
\alias{keep_pages}
\title{Keep only needed pages}
\usage{
keep_pages(pdf, pages)
}
\arguments{
\item{pdf}{pdf object}

\item{pages}{pages to keep}
}
\value{
pdf with only pages specified by pages
}
\description{
Keep only needed pages
}
\examples{
p <- pdf_load(test_pdf_text(pages = "all"))
p <- keep_pages(p, seq(7, 56))
}
