% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{make_row}
\alias{make_row}
\title{Make filler row}
\usage{
make_row(cols, row_len)
}
\arguments{
\item{cols}{location of col breaks}

\item{row_len}{total length of row}
}
\value{
character of length 1, filler row

pdf2tab:::make_row(c(2, 4), 6)
}
\description{
Make filler row
}
