% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_load_funs.R
\name{pdf_load}
\alias{pdf_load}
\title{Load PDF}
\usage{
pdf_load(pdf_file, cols = NULL, rowss = NULL, pages = NULL,
  drop_lines = c(0, 0), ...)
}
\arguments{
\item{...}{s}

\item{filename}{s}

\item{text}{s}
}
\value{
s
}
\description{
Load PDF
}
\examples{
pdf_load(test_pdf_text())
}
