% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{pdf_page}
\alias{pdf_page}
\title{Create a PDF Page}
\usage{
pdf_page(text, rows = NULL, cols = NULL)
}
\arguments{
\item{text}{Text of page}

\item{rows}{numeric, rows preceding row breaks}

\item{cols}{numeric, characters preceding column breaks}
}
\value{
a pdf_page object

str(pdf2tab:::pdf_page(list("HEADER 1         HEADER 2         HEADER 3",
         " adsdsf           asd             asdfad")))
}
\description{
Create a PDF Page
}
