% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{print.pdf_page}
\alias{print.pdf_page}
\title{Print a PDF page}
\usage{
\method{print}{pdf_page}(x, ...)
}
\arguments{
\item{x}{pdf_page}

\item{...}{nothing right now}
}
\value{
x (invisibly)
}
\description{
Print a PDF page
}
\examples{
pdf_page(list("a   b   c", "1   2   3"))
pdf_page(list("ABCDEF", "123456"), cols = c(2, 4))
pdf_page(list("ABCDEF", "ABCDEF", "ABCDEF", "123456"), rows = 3)
pdf_page(list("ABCDEF", "ABCDEF", "ABCDEF", "123456"),
cols = c(2, 4), rows = c(3, 4))
}
