% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentencing_pdf.R
\docType{data}
\name{sentencing_pdf}
\alias{sentencing_pdf}
\title{USSC Sentencing Databook}
\format{An 86 element character vector, each element is one page from the pdf.}
\source{
\url{https://www.ussc.gov/research/datafiles/commission-datafiles}
}
\usage{
sentencing_pdf
}
\description{
This is the result of running pdftools::pdf_text on the file in
`extdata/USSC_Public_Release_Codebook_FY99_FY16.pdf``
}
\keyword{datasets}
