% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{set_cols.pdf_page}
\alias{set_cols.pdf_page}
\title{Set Cols of PDF Page}
\usage{
\method{set_cols}{pdf_page}(x, val)
}
\arguments{
\item{x}{page of pdf}

\item{val}{cols, numeric}
}
\value{
x with cols set
}
\description{
Set Cols of PDF Page
}
