% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{set_page_rows}
\alias{set_page_rows}
\title{Set Rows for Page}
\usage{
set_page_rows(x, val)
}
\arguments{
\item{x}{pdf_age}

\item{val}{value}
}
\value{
pdf_page
}
\description{
Set Rows for Page
}
\examples{
p <- pdf_page("ABC", 1, 2)
set_page_rows(p, 3)
}
