% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{set_rows.pdf_page}
\alias{set_rows.pdf_page}
\title{Set Rows of PDF Page}
\usage{
\method{set_rows}{pdf_page}(x, val)
}
\arguments{
\item{x}{page of pdf}

\item{val}{rows, numeric}
}
\value{
x with rows set
}
\description{
Set Rows of PDF Page
}
