% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_funs.R
\name{split_text}
\alias{split_text}
\title{Title}
\usage{
split_text(pdf, by_char = "\\n")
}
\arguments{
\item{by_char}{}
}
\description{
Title
}
\examples{
pdf(c("ABC\\nDEF\\nGHI", "abc\\ndef\\nghi")) \%>\% split_text()
}
