% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_load_funs.R
\name{test_pdf_text}
\alias{test_pdf_text}
\title{Load the text of a pdf for testing}
\usage{
test_pdf_text(which = "sentencing", pages = NULL)
}
\arguments{
\item{which}{which pdf, currently just supports default}

\item{pages}{which pages, table starts on page 7}
}
\value{
list, each element 1 page
}
\description{
Load the text of a pdf for testing
}
\examples{
test_pdf_text()
}
