% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{add_print_cols}
\alias{add_print_cols}
\title{Add column breaks}
\usage{
add_print_cols(x, cols)
}
\arguments{
\item{x}{list of characters}

\item{cols}{column breaks}
}
\value{
x with column break characters added

t <- pdf2tab:::pdf_page(list("123456789123456789",
         "123456789123456789"),
         cols = c(3, 6, 15))
pdf2tab:::add_print_cols(pdf2tab:::get_page_text(t),
pdf2tab:::get_page_cols(t))
pdf2tab:::add_print_cols(pdf2tab:::get_page_text(t), NULL)
}
\description{
Add column breaks
}
