% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{get_page_rows}
\alias{get_page_rows}
\title{Get PDF Page Rows}
\usage{
get_page_rows(x)
}
\arguments{
\item{x}{pdf_age}
}
\value{
numeric, rows of x

p <- pdf2tab:::pdf_page("ABC", 1, 2)
pdf2tab:::get_page_rows(p)
}
\description{
Get PDF Page Rows
}
