% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{get_page_text}
\alias{get_page_text}
\title{Get Text from Page}
\usage{
get_page_text(x)
}
\arguments{
\item{x}{pdf_age}
}
\value{
character, text of x
}
\description{
Get Text from Page
}
\examples{
p <- pdf_page("ABC", 1, 2)
get_page_text(p)
}
