% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_funs.R
\name{pdf_load}
\alias{pdf_load}
\title{Load PDF}
\usage{
pdf_load(filename = NULL, n_col = NULL, text = NULL, pages = NULL,
  drop_from_top = 0, drop_from_bottom = 0, ...)
}
\arguments{
\item{filename}{s}

\item{text}{s}

\item{...}{s}
}
\value{
s
}
\description{
Load PDF
}
\examples{
pdf_load(text = test_pdf_text())
}
