% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_load_funs.R
\name{pdf_load}
\alias{pdf_load}
\title{Load PDF}
\usage{
pdf_load(pdf_file, cols = NULL, rows = NULL, pages = NULL,
  drop_lines = c(NULL, NULL), ...)
}
\arguments{
\item{...}{s}

\item{filename}{s}

\item{text}{s}
}
\value{
s
}
\description{
Load PDF
}
\examples{
p <- pdf_load(test_pdf_text())
p <- pdf_load(test_pdf_text(), drop_lines = c(1, 2))
p <- pdf_load(test_pdf_text(pages = "all"), pages = seq(7, 45))
}
