% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_load_funs.R
\name{pdf_load}
\alias{pdf_load}
\title{Load PDF}
\usage{
pdf_load(pdf_file)
}
\arguments{
\item{pdf_file}{a filename of a pdf or a list of characters}

\item{...}{optional arguments passed to pdftools::pdf_text}
}
\value{
a pdf object
}
\description{
Load PDF
}
\examples{
p <- pdf_load()
}
