% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_funs.R
\name{pdf_load}
\alias{pdf_load}
\title{Title}
\usage{
pdf_load(filename = NULL, text = NULL, pages = NULL, drop_from_top = 0,
  drop_from_bottom = 0, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
#pdf_load(text = c("ABC\\nDEF\\nGHI", "abc\\ndef\\nghi"))
}
