% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{set_page_attr}
\alias{set_page_attr}
\title{Set Attribute for Page}
\usage{
set_page_attr(x, which, val)
}
\arguments{
\item{x}{pdf_age}

\item{which}{attribute}

\item{val}{value}
}
\value{
attribute of x

p <- pdf2tab:::pdf_page("ABC", 1, 2)
pdf2tab:::set_page_attr(p, "text", "DEF")
}
\description{
Set Attribute for Page
}
