% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_page_funs.R
\name{set_page_text}
\alias{set_page_text}
\title{Set Text for Page}
\usage{
set_page_text(x, val)
}
\arguments{
\item{x}{pdf_age}

\item{val}{value}
}
\value{
pdf_page

p <- pdf2tab:::pdf_page("ABC", 1, 2)
pdf2tab:::set_page_text(p, "DEF")
}
\description{
Set Text for Page
}
