% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_funs.R
\name{test_pdf_text}
\alias{test_pdf_text}
\title{Sample PDF Text for Testing}
\usage{
test_pdf_text(which = "sentencing")
}
\value{
list of characters
}
\description{
Sample PDF Text for Testing
}
\examples{
test_pdf_text()
}
